+++ scout/constants/igv_tracks.py
@@ -13,10 +13,8 @@ HG38CYTOBAND_URL = (
     "https://s3.amazonaws.com/igv.broadinstitute.org/annotations/hg38/cytoBandIdeo.txt"
 )
 
-HG38GENES_FORMAT = "gtf"
-HG38GENES_URL = "https://s3.amazonaws.com/igv.broadinstitute.org/annotations/hg38/genes/Homo_sapiens.GRCh38.80.sorted.gtf.gz"
-HG38GENES_INDEX_URL = "https://s3.amazonaws.com/igv.broadinstitute.org/annotations/hg38/genes/Homo_sapiens.GRCh38.80.sorted.gtf.gz.tbi"
-HG19GENES_FORMAT = "refgene"
+HG38GENES_URL = "https://s3.amazonaws.com/igv.org.genomes/hg38/refGene.txt.gz"
+
 HG19GENES_URL = "https://s3.amazonaws.com/igv.org.genomes/hg19/ncbiRefSeq.sorted.txt.gz"
 HG19GENES_INDEX_URL = "https://s3.amazonaws.com/igv.org.genomes/hg19/ncbiRefSeq.sorted.txt.gz.tbi"
 
@@ -117,14 +115,8 @@ HUMAN_GENES_37 = {
 # Human genes track, build 38
 HUMAN_GENES_38 = {
     "name": "Genes",
-    "track_name": "genes_track",
-    "type": "annotation",
-    "sourceType": "file",
-    "displayMode": "EXPANDED",
-    "visibilityWindow": 300000000,
-    "format": HG38GENES_FORMAT,
     "url": HG38GENES_URL,
-    "indexURL": HG38GENES_INDEX_URL,
+    "order": 1000000,
 }
 
 CASE_SPECIFIC_TRACKS = {